clients: Update links for Cantata and QMPDClient
Looking at the activity of the Cantata client, it may very well be worth
moving back to the list of maintained clients.

The URL for QMPDClient seemed not to be reachable anymore, but a copy of the
source was found on the client author's Gitlab page. This commit updates the
URL to point to that page.