MPDroid doesn't work on new phones
I moved it to the unmaintained section because the author said in 2018 that it is unmaintained: abarisain/dmix#856 (comment)

The app has been working, but it can't be installed on phones with newer processors (side loading the APK gives the generic You can't install the app on your device error, and F-Droid doesn't offer it for installation). On older phones, it mostly works but has some significant bugs.

The last work on the repository was 4 years ago and there are 150+ issues open. Several of those issues are about the major work that would be needed to make the app compatible with newer Android devices/OS versions. It is clearly not going to happen.