Issue 156: add missing packages to instructions
I was not able to build or view the site when using the instructions in a
minimal Ruby environment.  More details in the issue, but I found that I
needed to install builder in order to build the site, and puma in order to
view it.

This updates the instructions to install both packages.